Restructured test structure
