+++ tests/store/handlers/test_readhandler.py
@@ -1,9 +1,34 @@
-"""Tests for the store api"""
-import datetime
+"""Tests for finding tags in store."""
+from datetime import datetime
 from typing import List
-from housekeeper.store.api.core import Store
 
-from housekeeper.store.models import File
+from housekeeper.store import Store
+from housekeeper.store.models import Tag, File
+
+
+def test_tag_with_tag_name(populated_store: Store, sample_tag_name: str):
+    """Test fetching a tag from database given a tag name."""
+    # GIVEN a populated store and a tag name
+
+    # WHEN retrieving a tag from Store
+    test_tag = populated_store.get_tag(tag_name=sample_tag_name)
+
+    # THEN a tag should be returned
+    assert isinstance(test_tag, Tag)
+
+    # THEN the tag should have the tag name gotten as parameter
+    assert test_tag.name == sample_tag_name
+
+
+def test_tag_without_tag_name(populated_store: Store):
+    """Test fetching a tag from database without specifying a tag name returns None."""
+    # GIVEN a populated store
+
+    # WHEN retrieving a tag from Store
+    test_tag = populated_store.get_tag()
+
+    # THEN a tag should be returned
+    assert test_tag is None
 
 
 def test_get_files_before(populated_store, bundle_data_old, time_stamp_now):
